Clone and Default implementations reorganized
